Temporary commit to move code to different PC

+++ benches/sort_bench.rs
@@ -19,7 +19,7 @@ fn quicksort_benchmark(c: &mut Criterion) {
     bench_sort_func(c, "quicksort", quick_sort);
 }
 
-fn bench_sort_func<F>(c: &mut Criterion, group_name: &str, mut f: F) where F: FnMut(&mut [i64]) {
+fn bench_sort_func<F>(c: &mut Criterion, group_name: &str, mut f: F) where F: FnMut(&mut [i32]) {
     let mut group = c.benchmark_group(group_name);
     for n in (1..4).map(|i| 10_usize.pow(i)) {
         group.bench_with_input(BenchmarkId::from_parameter(n), &n, |b, &n| {
@@ -30,8 +30,8 @@ fn bench_sort_func<F>(c: &mut Criterion, group_name: &str, mut f: F) where F: Fn
     }
 }
 
-fn random_array(n: usize) -> Vec<i64>{
-    thread_rng().gen_iter::<i64>().take(n).collect()
+fn random_array(n: usize) -> Vec<i32>{
+    thread_rng().gen_iter::<i32>().take(n).collect()
 }
 
 criterion_group!(
